refactor: renamed queries
