+++ src/schema/query/apiConfiguration.query.ts
