+++ src/schema/query/apiConfigurations.query.ts
