+++ src/schema/query/application.query.ts
