+++ src/schema/query/applications.query.ts
