+++ src/schema/query/index.ts
@@ -1,40 +1,40 @@
 import { GraphQLObjectType } from 'graphql';
-import resources from './resources';
-import resource from './resource';
-import notifications from './notifications';
-import forms from './forms';
-import form from './form';
-import records from './records';
-import record from './record';
-import recordsAggregation from './recordsAggregation';
-import dashboards from './dashboards';
-import dashboard from './dashboard';
-import users from './users';
-import me from './me';
-import role from './role';
-import roles from './roles';
-import rolesFromApplications from './rolesFromApplications';
-import step from './step';
-import steps from './steps';
-import workflow from './workflow';
-import workflows from './workflows';
-import page from './page';
-import pages from './pages';
-import application from './application';
-import applications from './applications';
-import permissions from './permissions';
-import channels from './channels';
-import positionAttributes from './positionAttributes';
-import apiConfiguration from './apiConfiguration';
-import apiConfigurations from './apiConfigurations';
-import pullJobs from './pullJobs';
-import referenceData from './referenceData';
-import referenceDatas from './referenceDatas';
-import recordHistory from './recordHistory';
-import setting from './setting';
-import user from './user';
-import group from './group';
-import groups from './groups';
+import resources from './resources.query';
+import resource from './resource.query';
+import notifications from './notifications.query';
+import forms from './forms.query';
+import form from './form.query';
+import records from './records.query';
+import record from './record.query';
+import recordsAggregation from './recordsAggregation.query';
+import dashboards from './dashboards.query';
+import dashboard from './dashboard.query';
+import users from './users.query';
+import me from './me.query';
+import role from './role.query';
+import roles from './roles.query';
+import rolesFromApplications from './rolesFromApplications.query';
+import step from './step.query';
+import steps from './steps.query';
+import workflow from './workflow.query';
+import workflows from './workflows.query';
+import page from './page.query';
+import pages from './pages.query';
+import application from './application.query';
+import applications from './applications.query';
+import permissions from './permissions.query';
+import channels from './channels.query';
+import positionAttributes from './positionAttributes.query';
+import apiConfiguration from './apiConfiguration.query';
+import apiConfigurations from './apiConfigurations.query';
+import pullJobs from './pullJobs.query';
+import referenceData from './referenceData.query';
+import referenceDatas from './referenceDatas.query';
+import recordHistory from './recordHistory.query';
+import setting from './setting.query';
+import user from './user.query';
+import group from './group.query';
+import groups from './groups.query';
 
 /** GraphQL query type definition */
 const Query = new GraphQLObjectType({
