+++ src/schema/query/notifications.query.ts
