+++ src/schema/query/positionAttributes.query.ts
