+++ src/schema/query/resource.query.ts
