+++ src/schema/query/rolesFromApplications.query.ts
