+++ src/schema/query/setting.query.ts
